Add files via upload
Changed webpage_server.py and main.py to work on a linux system had to change webpage directory as apache uses a directory and permissions and starting apache is different in linux. Script is now designed to be ran with sudo infront as otherwise there will be permission errors from writing files in apache directories and using the systemctl bash command.